% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cevodata.R
\docType{data}
\name{variant_classification}
\alias{variant_classification}
\title{Variant_Classification to IMPACT table used in the package}
\description{
Variant_Classification to IMPACT table used in the package
}
